groot/rtree: lift WriteVar out of Writer
